+++ apps/dialtone-documentation/docs/.vuepress/baseComponents/tokens/utilities.js
@@ -73,8 +73,13 @@ const addTokensToCategories = (token, format, structure) => {
   }
 
   if ((CATEGORY_MAP.component.includes(splitKeys[0]) && (splitKeys[1] === 'color')) ||
-    (CATEGORY_MAP.component.includes(splitKeys[1]) && splitKeys[0] === 'theme')) {
+    (CATEGORY_MAP.component.includes(splitKeys[1]) && splitKeys[0] === 'shell')) {
     structure.color.components._children.push(displayToken);
+    return;
+  }
+
+  if (key.startsWith('color/chart')) {
+    structure.color.chart._children.push(displayToken);
     return;
   }