+++ packages/dialtone-tokens/build-ios.js
@@ -4,27 +4,19 @@
  * Builds the ios tokens to the correct final package format (output to dist_ios)
  */
 import fs from 'fs';
+import { glob } from 'glob';
 
-const THEMES = [
-  'base-light',
-  'base-dark',
-  'dp-light',
-  'dp-dark',
-  'tmo-light',
-  'tmo-dark',
-  'expressive-light',
-  'expressive-dark',
-  'expressive-sm-light',
-  'expressive-sm-dark',
-];
+const THEMES = glob.sync('dist/ios/tokens-*.swift');
 
 if (!fs.existsSync('dist_ios/Sources/DialtoneTokens')) {
   fs.mkdirSync('dist_ios/Sources/DialtoneTokens', { recursive: true });
 }
 
-THEMES.forEach(theme => {
-  fs.copyFile(`dist/ios/tokens-${theme}.swift`, `dist_ios/Sources/DialtoneTokens/tokens-${theme}.swift`, (err) => {
-    if (err) throw err;
-    console.log(`dialtone.swift was copied to dist_ios/Sources/DialtoneTokens/tokens-${theme}.swift`);
+THEMES
+  .map(theme => theme.replace('dist/ios/', ''))
+  .forEach(themeName => {
+    fs.copyFile(`dist/ios/${themeName}`, `dist_ios/Sources/DialtoneTokens/${themeName}`, (err) => {
+      if (err) throw err;
+      console.log(`Copied ${themeName} to dist_ios`);
+    });
   });
-});