+++ packages/dialtone-vue2/recipes/conversation_view/message_input/message_input_default.story.vue
@@ -54,6 +54,7 @@
         @paste-media="$attrs.onPasteMedia"
         @notice-close="$attrs.onNoticeClose"
         @cancel="$attrs.onCancel"
+        @emoji-scroll-bottom-reached="$attrs.onEmojiScrollBottomReached"
       >
         <template
           v-if="$attrs.emojiGiphyPicker"