+++ packages/dialtone-vue3/components/toggle/toggle.vue
@@ -162,7 +162,7 @@ export default {
     inputListeners () {
       return {
         ...removeClassStyleAttrs(this.$attrs),
-        onClick: _ => this.toggleCheckedValue(),
+        onClick: () => this.toggleCheckedValue(),
       };
     },