+++ packages/eslint-plugin-dialtone/tests/lib/rules/deprecated-base-color-classes.js
@@ -12,7 +12,7 @@ const RuleTester = require('eslint').RuleTester;
 // ------------------------------------------------------------------------------
 
 const ruleTester = new RuleTester({
-  // eslint-disable-next-line node/no-extraneous-require
+  // eslint-disable-next-line n/no-extraneous-require
   parser: require.resolve('vue-eslint-parser'),
   parserOptions: { ecmaVersion: 'latest' },
 });